Added test functions to one file
